+++ CHANGELOG.md
@@ -8,6 +8,7 @@ in 1.x and 0.x versions.
 
   * Added cutlist same size part grouping option
   * Added cutlist export source option
+  * Added cutlist hide entity names option
   * Added cutlist real area for sheet good part
   * Added cutlist auto orient display
   * Added cutlist part orientation check